requirements-tests.txt
flake8
pep8-naming
pytest
selenium
pytest-flask
